memperbaiki untuk navigate ke RincianDeposit dan refactor bebarapa code di deposit dan adapternya

+++ app/src/main/java/com/example/konsinyasiapp/entities/ProductInDeposit.kt
@@ -24,6 +24,6 @@ class ProductInDeposit(
     var jumlahQuantity: String,
 
     @ColumnInfo(name = "return_quantity")
-    var returnQuantity: Int
+    var returnQuantity: Long
 
 ) : Parcelable
